+++ backend/src/controllers/reports/Tab_48_rpt_PF_FinanceRecoverySummary.js
@@ -4,10 +4,11 @@ const model = require("@models/reports/Tab_48_rpt_PF_FinanceRecoverySummary");
 const utils = require("./helpers");
 const what = { single: "report", plural: "reports" };
 const controller = useController(model, what);
+const _ = require("lodash");
 
 // Template and data reading
 const cdogs = useCommonComponents("cdogs");
-const { getReport, getDocumentApiBody, pdfConfig } = utils;
+const { getReport, getDocumentApiBody, pdfConfig, groupByProperty } = utils;
 controller.getReport = getReport;
 
 /**
@@ -20,19 +21,26 @@ controller.getReport = getReport;
 controller.Tab_48_rpt_PF_FinanceRecoverySummary = async (request, reply) => {
   controller.userRequires(request, "PMO-Reports-Capability", reply);
   try {
-    // Get the data from the database.
+    // gets the data from the database
     const getDate = async () => new Date();
+    const [{ fiscal_year }] = await model.getFiscalYear(request.query);
     const report = await model.Tab_48_rpt_PF_FinanceRecoverySummary(request.query);
     const report_totals = await model.Tab_48_totals(request.query);
     const report_grand_totals = await model.Tab_48_grand_totals(request.query);
-    const [{ fiscal_year }] = await model.getFiscalYear(request.query);
+
+    // shape the dataset so it can be parsed by the templating engine properly
+    const reportByPortfolio = groupByProperty(report, "portfolio_name");
+    const totalsByPortfolio = _.keyBy(report_totals, "portfolio_name");
+    const reportsByPortfolioWithTotals = _.map(reportByPortfolio, (portfolio) => ({
+      ...portfolio,
+      portfolio_totals: totalsByPortfolio[portfolio.portfolio_name],
+    }));
 
     const result = {
       date: await getDate(),
       fiscal: fiscal_year,
-      report,
-      report_totals,
-      report_grand_totals,
+      report: reportsByPortfolioWithTotals,
+      grand_totals: _.first(report_grand_totals),
     };
 
     const body = await getDocumentApiBody(result, "Tab_48_rpt_PF_FinanceRecoverySummary.docx");